# imaxx-smart-office-be-dev/docs/swagger.yaml
basePath: /api
definitions:
  controller.EditPassword:
    properties:
      password:
        type: string
    required:
    - password
    type: object
...
  request.AuthenticateRequest:
    properties:
      password:
        maxLength: 255
        minLength: 1
        type: string
      username:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - password
    - username
    type: object
...
  request.CreateUserRequest:
    properties:
      email:
        maxLength: 255
        minLength: 1
        type: string
      employed_at:
        type: string
      first_name:
        maxLength: 255
        minLength: 1
        type: string
      last_name:
        maxLength: 255
        minLength: 1
        type: string
      password:
        maxLength: 255
        minLength: 1
        type: string
      role_ids:
        items:
          type: string
        type: array
      status:
        type: integer
      username:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - email
    - employed_at
    - first_name
    - last_name
    - password
    - role_ids
    - status
    - username
    type: object
...
  request.UpdateUserRequest:
    properties:
      email:
        maxLength: 255
        minLength: 1
        type: string
      employed_at:
        type: string
      first_name:
        maxLength: 255
        minLength: 1
        type: string
      last_name:
        maxLength: 255
        minLength: 1
        type: string
      password:
        type: string
      role_ids:
        items:
          type: string
        type: array
      status:
        type: integer
      username:
        type: string
    required:
    - email
    - employed_at
    - first_name
    - last_name
    - role_ids
    - status
    type: object
...
  response.DefaultResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:5000
info:
  contact: {}
  description: iMaxx Office API with Swagger
  title: iMaxx Office API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: find all user
      parameters:
      - default: 1
        description: page number
        in: query
        name: page
        type: integer
      - default: 10
        description: limit per page number
        in: query
        name: limit
        type: integer
      - default: created_at desc
        description: sort sql string ex. created_at desc
        in: query
        name: sort
        type: string
      - description: keyword search
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindAllUser user
      tags:
      - users
    post:
      consumes:
      - application/json
      description: create user
      parameters:
      - description: create user payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: CreateUser user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: delete user
      parameters:
      - description: User ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: DeleteUser user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: find by id user
      parameters:
      - description: User ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindById user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: update user
      parameters:
      - description: update user payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateUserRequest'
      - description: User ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: UpdateUser user
      tags:
      - users
  /users/active:
    get:
      consumes:
      - application/json
      description: find active all user
      parameters:
      - default: 1
        description: page number
        in: query
        name: page
        type: integer
      - default: 10
        description: limit per page number
        in: query
        name: limit
        type: integer
      - default: created_at desc
        description: sort sql string ex. created_at desc
        in: query
        name: sort
        type: string
      - description: keyword search
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindActiveAllUser user
      tags:
      - users
  /users/authenticate:
    post:
      consumes:
      - application/json
      description: authenticate
      parameters:
      - description: authenticate payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.AuthenticateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      summary: Authenticate user
      tags:
      - users
  /users/me:
    get:
      consumes:
      - application/json
      description: Get current user profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Get my profile users
      tags:
      - users
  /users/me/password:
    patch:
      consumes:
      - application/json
      description: update password by user
      parameters:
      - description: update user password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.EditPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Update user password
      tags:
      - users
  /users/me/upload-profile:
    patch:
      consumes:
      - multipart/form-data
      description: upload user avatar file
      parameters:
      - description: upload avatar file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Upload image for user avatar
      tags:
      - users
  /users/refresh-token:
    post:
      consumes:
      - application/json
      description: refresh token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: RefreshToken user
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer ' followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"